requirements.txt
torch
torchvision
numpy
pandas
pyarrow
matplotlib
seaborn
tqdm
pyyaml
tensorboard
wandb
pytest
